Implement full featured PAM interface
There once was a python package "PAM" which provided a very nice and
detailed implementation of pam functionality.
That project seems dead: http://www.pangalactic.org/PyPAM

The package also doesn't provide a Python 3 version in Debian Buster,
and maybe won't in Debian Bullseye (while Ubuntu already migrated it to
Python 3):
https://packages.debian.org/source/buster/python-pam

The idea here is to provide a feature complete full compliant replacement.
On top of this, every current functionality of python-pampy is added
using this new version.