internals: Root conversation and closure data
If the conversation and closure data are not rooted to the class
then they can be garbage collected before the conversation function
is invoked.

That leads to a crash in pam_lastlog during session open time.

This commit addressed the issue by assigning the state to the class
as properties.