Don't ship pam.py in the module
The comment at the top of this file says:
"This is a legacy file, it is not used. Here for example."
so it seems sensible to no longer ship it as part of the pam
module itself. This moves it to an examples/ directory which
will not be included in builds of the actual module, and updates
the comment.

Signed-off-by: Adam Williamson <awilliam@redhat.com>